refactor: 패키지 구조, 네이밍 변경 (common -> match)
