+++ match-api/src/main/java/com/stocat/match/MatchApiApplication.java
@@ -1,0 +1,14 @@
+package com.stocat.match;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootApplication
+public class MatchApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MatchApiApplication.class, args);
+    }
+
+}
