+++ match-api/src/main/resources/application.yml
@@ -8,6 +8,15 @@ spring:
       host: localhost
       port: 6379
 
+# Redis Stream 설정
+redis:
+  stream:
+    key-prefix: "orderbook:stream:"
+    consumer-group: "matching-engine"
+    block-timeout: 5s
+    batch-size: 10
+    symbols: []  # 초기화할 심볼 목록 (필요 시 설정)
+
 # KIS API 설정 (실전투자)
 kis:
   websocket:
