+++ match-api/src/test/java/com/stocat/match/MatchApiApplicationTests.java
@@ -1,0 +1,13 @@
+package com.stocat.match;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MatchApiApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
