--- match-api/src/test/java/com/stocat/matchapi/MatchApiApplicationTests.java
@@ -1,13 +1,0 @@
-package com.stocat.matchapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MatchApiApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
