+++ match-domain/src/main/java/com/stocat/match/domain/AssetsCategory.java
@@ -1,0 +1,5 @@
+package com.stocat.match.domain;
+
+public enum AssetsCategory {
+    KOR_STOCK, US_STOCK, CRYPTO, USD, KRW
+}
