+++ match-domain/src/main/java/com/stocat/match/domain/Currency.java
@@ -1,0 +1,18 @@
+package com.stocat.match.domain;
+
+import lombok.Getter;
+
+@Getter
+public enum Currency {
+    KRW(AssetsCategory.KRW),
+    USD(AssetsCategory.USD),
+    BTC(null),
+    USDT(null),
+    ;
+
+    private final AssetsCategory category;
+
+    Currency(AssetsCategory category) {
+        this.category = category;
+    }
+}
