+++ match-domain/src/main/java/com/stocat/match/domain/TradeSide.java
@@ -1,0 +1,9 @@
+package com.stocat.match.domain;
+
+/**
+ * 매매 구분
+ */
+public enum TradeSide {
+    BUY,
+    SELL;
+}
