+++ match-domain/src/main/java/com/stocat/match/domain/order/OrderTif.java
@@ -1,0 +1,14 @@
+package com.stocat.match.domain.order;
+
+import lombok.Getter;
+
+@Getter
+public enum OrderTif {
+    GTC("취소 전까지 유효"), IOC("미체결 수량 자동 취소");
+
+    private final String description;
+
+    OrderTif(String description) {
+        this.description = description;
+    }
+}
