--- match-domain/src/main/java/com/stocat/matchdomain/AssetsCategory.java
@@ -1,5 +1,0 @@
-package com.stocat.matchdomain;
-
-public enum AssetsCategory {
-    KOR_STOCK, US_STOCK, CRYPTO, USD, KRW
-}
