--- match-domain/src/main/java/com/stocat/matchdomain/Currency.java
@@ -1,18 +1,0 @@
-package com.stocat.matchdomain;
-
-import lombok.Getter;
-
-@Getter
-public enum Currency {
-    KRW(AssetsCategory.KRW),
-    USD(AssetsCategory.USD),
-    BTC(null),
-    USDT(null),
-    ;
-
-    private final AssetsCategory category;
-
-    Currency(AssetsCategory category) {
-        this.category = category;
-    }
-}
