--- match-domain/src/main/java/com/stocat/matchdomain/TradeSide.java
@@ -1,9 +1,0 @@
-package com.stocat.matchdomain;
-
-/**
- * 매매 구분
- */
-public enum TradeSide {
-    BUY,
-    SELL;
-}
