--- match-domain/src/main/java/com/stocat/matchdomain/fill/Fill.java
@@ -1,16 +1,0 @@
-package com.stocat.matchdomain.fill;
-
-import com.stocat.matchdomain.TradeSide;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-
-public record Fill(
-        Long orderId,
-        String symbol,
-        TradeSide side,
-        BigDecimal price,
-        BigDecimal quantity,
-        LocalDateTime filledAt
-) {
-}
