--- match-domain/src/main/java/com/stocat/matchdomain/order/OrderTif.java
@@ -1,14 +1,0 @@
-package com.stocat.matchdomain.order;
-
-import lombok.Getter;
-
-@Getter
-public enum OrderTif {
-    GTC("취소 전까지 유효"), IOC("미체결 수량 자동 취소");
-
-    private final String description;
-
-    OrderTif(String description) {
-        this.description = description;
-    }
-}
