--- match-domain/src/main/java/com/stocat/matchdomain/order/OrderType.java
@@ -1,14 +1,0 @@
-package com.stocat.matchdomain.order;
-
-import lombok.Getter;
-
-@Getter
-public enum OrderType {
-    LIMIT("지정가"), MARKET("시장가");
-
-    private final String description;
-
-    OrderType(String description) {
-        this.description = description;
-    }
-}
