+++ match-mysql/src/main/java/com/stocat/match/config/MysqlJpaConfig.java
@@ -1,0 +1,15 @@
+package com.stocat.match.config;
+
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
+
+@Configuration
+@EnableTransactionManagement
+@EnableJpaRepositories(basePackages = "com.stocat")
+@EntityScan(basePackages = "com.stocat")
+@EnableJpaAuditing
+public class MysqlJpaConfig {
+}
