--- match-redis/src/test/java/com/stocat/common/redis/StocatRedisApplicationTests.java
@@ -1,14 +1,0 @@
-package com.stocat.common.redis;
-
-import com.stocat.common.redis.config.RedisConfig;
-import org.junit.jupiter.api.Test;
-import org.springframework.test.context.ContextConfiguration;
-
-@ContextConfiguration(classes = {RedisConfig.class})
-class StocatRedisApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
