+++ match-redis/src/test/java/com/stocat/match/redis/StocatRedisApplicationTests.java
@@ -1,0 +1,14 @@
+package com.stocat.match.redis;
+
+import com.stocat.match.redis.config.RedisConfig;
+import org.junit.jupiter.api.Test;
+import org.springframework.test.context.ContextConfiguration;
+
+@ContextConfiguration(classes = {RedisConfig.class})
+class StocatRedisApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
